data_engineering_project/End-to-End Data Engineering Project using AWS + Apache Iceberg + Snowflake/requirements.txt
boto3
pandas
pyarrow
snowflake-connector-python
snowflake-snowpark-python
